+++ core-ng/src/main/java/core/framework/impl/validate/FieldValidator.java
@@ -4,5 +4,5 @@ package core.framework.impl.validate;
  * @author neo
  */
 interface FieldValidator {
-    void validate(Object instance, ValidationErrors errors);
+    void validate(Object instance, ValidationErrors errors, boolean partial);
 }