+++ core-ng/src/main/java/core/framework/impl/validate/ListValidator.java
@@ -15,9 +15,9 @@ public class ListValidator implements FieldValidator {
     }
 
     @Override
-    public void validate(Object list, ValidationErrors errors) {
+    public void validate(Object list, ValidationErrors errors, boolean partial) {
         if (list instanceof List) {
-            ((List<?>) list).forEach(value -> valueValidator.validate(value, errors));
+            ((List<?>) list).forEach(value -> valueValidator.validate(value, errors, partial));
         } else if (list != null) {
             throw Exceptions.error("expected list, actualClass={}", list.getClass().getCanonicalName());
         }