+++ core-ng/src/main/java/core/framework/impl/validate/MapValidator.java
@@ -15,9 +15,9 @@ public class MapValidator implements FieldValidator {
     }
 
     @Override
-    public void validate(Object map, ValidationErrors errors) {
+    public void validate(Object map, ValidationErrors errors, boolean partial) {
         if (map instanceof Map) {
-            ((Map<?, ?>) map).forEach((key, value) -> valueValidator.validate(value, errors));
+            ((Map<?, ?>) map).forEach((key, value) -> valueValidator.validate(value, errors, partial));
         } else if (map != null) {
             throw Exceptions.error("expected map, actualClass={}", map.getClass().getCanonicalName());
         }