+++ core-ng/src/main/java/core/framework/impl/validate/MinValidator.java
@@ -16,7 +16,7 @@ public class MinValidator implements FieldValidator {
     }
 
     @Override
-    public void validate(Object value, ValidationErrors errors) {
+    public void validate(Object value, ValidationErrors errors, boolean partial) {
         if (value == null) return;
 
         if (value instanceof Number) {