+++ core-ng/src/test/java/core/framework/impl/db/EntityWithAssignedId.java
@@ -4,6 +4,7 @@ import core.framework.api.db.Column;
 import core.framework.api.db.PrimaryKey;
 import core.framework.api.db.Table;
 import core.framework.api.validate.Length;
+import core.framework.api.validate.NotNull;
 
 import java.math.BigDecimal;
 
@@ -21,6 +22,7 @@ public class EntityWithAssignedId {
     @Column(name = "string_field")
     public String stringField;
 
+    @NotNull
     @Column(name = "int_field")
     public Integer intField;